User Logout Fixing Post Deletion

+++ app/Http/Controllers/PostController.php
@@ -3,13 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Post;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
+
 
 class PostController extends Controller
 {
 	 public function getDashboard()
 	{
-	  $posts = Post::all();
+	  $posts = Post::orderBy('created_at','desc')->get();
       return view('dashboard',['posts' => $posts]);
     }
 
@@ -31,7 +34,12 @@ class PostController extends Controller
 	public function getDeletePost($post_id)
 	{
 		$post = Post::where('id', $post_id)->first();
+		if (Auth::User() != $post->user) {
+
+			return redirect()->back();
+		}
 		$post->delete();
 		return redirect()->route('dashboard')->with(['message' => 'Successfully deleted!']);
 	}
 }
+
